debug
git-svn-id: https://cbilsvn.pmacs.upenn.edu/svn/gus/WDK/trunk@19822 4235af6a-31f9-0310-9ad3-ecf6348f2534

+++ Model/src/main/java/org/gusdb/wdk/model/fix/InvalidStepReporter.java
@@ -1,5 +1,6 @@
 package org.gusdb.wdk.model.fix;
 
+import java.math.BigDecimal;
 import java.sql.ResultSet;
 import java.sql.SQLException;
 import java.util.List;
@@ -126,8 +127,8 @@ public class InvalidStepReporter extends BaseCLI {
     List<Map<String,Object>> results = handler.getResults();
     for (Map<String,Object> row : results) {
       String proj = (String)row.get("PROJECT_ID");
-      Integer is_val = (Integer)row.get("IS_VALID");
-      Integer cnt = (Integer)row.get("COUNT");
+      BigDecimal is_val = (BigDecimal)row.get("IS_VALID");
+      BigDecimal cnt = (BigDecimal)row.get("COUNT");
       System.out.println(proj + "\t" + is_val + "\t" + cnt);
     } 
   }
